+++ src/main/java/de/maxhenkel/voicechat/event/VoiceChatDisconnectedEvent.java
@@ -1,0 +1,12 @@
+package de.maxhenkel.voicechat.event;
+
+import net.minecraftforge.eventbus.api.Event;
+
+public class VoiceChatDisconnectedEvent extends Event {
+
+    @Override
+    public boolean isCancelable() {
+        return false;
+    }
+
+}