+++ src/main/java/de/maxhenkel/voicechat/voice/client/MicThread.java
@@ -35,7 +35,9 @@ public class MicThread extends Thread {
                 Utils.sleep(10);
             } else {
                 MicrophoneActivationType type = Main.CLIENT_CONFIG.microphoneActivationType.get();
-                if (type.equals(MicrophoneActivationType.PTT)) {
+                if (Main.CLIENT_VOICE_EVENTS.getPlayerStateManager().isDisabled()) {
+                    Utils.sleep(10);
+                } else if (type.equals(MicrophoneActivationType.PTT)) {
                     ptt();
                 } else if (type.equals(MicrophoneActivationType.VOICE)) {
                     voice();
@@ -51,7 +53,7 @@ public class MicThread extends Thread {
     private void voice() {
         wasPTT = false;
 
-        if (client.isMuted()) {
+        if (Main.CLIENT_VOICE_EVENTS.getPlayerStateManager().isMuted()) {
             activating = false;
             if (mic.isActive()) {
                 mic.stop();